Major Update

+++ powerplant-coding-challenge/Features/ProductionPlanCommand.cs
@@ -5,7 +5,7 @@ namespace powerplant_coding_challenge.Features;
 
 public class ProductionPlanCommand : IRequest<List<ProductionPlanCommandResponse>>
 {
-    public double Load { get; set; }
+    public decimal Load { get; set; }
     public List<Powerplant> Powerplants { get; set; } = [];
     public Fuels Fuels { get; set; } = new();
 }
