+++ powerplant-coding-challenge/Interfaces/ICostCalculator.cs
@@ -4,5 +4,5 @@ namespace powerplant_coding_challenge.Interfaces;
 
 public interface ICostCalculator
 {
-    double CalculateCostPerMWh(Powerplant powerplant, Fuels fuels);
+    decimal CalculateCostPerMWh(Powerplant powerplant, Fuels fuels);
 }
