+++ powerplant-coding-challenge/Interfaces/IProductionCalculator.cs
@@ -4,5 +4,5 @@ namespace powerplant_coding_challenge.Interfaces;
 
 public interface IProductionCalculator
 {
-    double CalculateProduction(Powerplant powerplant, double load, double windPercentage);
+    decimal CalculateProduction(Powerplant powerplant, decimal load, decimal windPercentage);
 }
