+++ powerplant-coding-challenge/Models/Fuels.cs
@@ -5,14 +5,14 @@ namespace powerplant_coding_challenge.Models;
 public class Fuels
 {
     [JsonPropertyName("gas(euro/MWh)")]
-    public double Gas { get; set; }
+    public decimal Gas { get; set; }
 
     [JsonPropertyName("kerosine(euro/MWh)")]
-    public double Kerosine { get; set; }
+    public decimal Kerosine { get; set; }
 
     [JsonPropertyName("co2(euro/ton)")]
-    public double Co2 { get; set; }
+    public decimal Co2 { get; set; }
 
     [JsonPropertyName("wind(%)")]
-    public double Wind { get; set; }
+    public decimal Wind { get; set; }
 }
