+++ powerplant-coding-challenge/Models/Powerplant.cs
@@ -4,7 +4,7 @@ public class Powerplant
 {
     public string Name { get; set; } = string.Empty;
     public string Type { get; set; } = string.Empty;
-    public double Efficiency { get; set; }
-    public double Pmin { get; set; }
-    public double Pmax { get; set; }
+    public decimal Efficiency { get; set; }
+    public decimal Pmin { get; set; }
+    public decimal Pmax { get; set; }
 }
